% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{.prepareDataForGrowthCurve}
\alias{.prepareDataForGrowthCurve}
\title{Prepare Data for growth curve plotting}
\usage{
.prepareDataForGrowthCurve(data, x, y, grouping, color)
}
\arguments{
\item{data}{A data frame containing the data to plot.}

\item{x}{The x-axis variable. Usually the time in hours after starting the experiment.}

\item{y}{The y-axis variable. Usually the concentration of microbes per mL as a raw number}

\item{grouping}{The grouping variable. Usually the name of the strain or condition.}

\item{color}{The color variable. Usually the name of the strain or condition.}
}
\description{
Prepare Data for growth curve plotting
}
\keyword{internal}
