% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPfamAnnotationForPDBFilenames.R
\name{get_assembly_id_from_target}
\alias{get_assembly_id_from_target}
\title{Extract PDB Assembly ID from Target Name}
\usage{
get_assembly_id_from_target(target)
}
\arguments{
\item{target}{A character string containing a PDB target name with assembly information.}
}
\value{
A data frame with columns:
\itemize{
\item \code{target}: The original target string.
\item \code{assembly_id}: The extracted PDB assembly ID.
}
}
\description{
This function extracts the PDB assembly ID from a given target string.
}
\keyword{:}
\keyword{internal}
