% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPfamAnnotationForPDBFilenames.R
\name{get_polymer_info}
\alias{get_polymer_info}
\title{Retrieve Polymer Entity Identifiers and Pfam Annotations from RCSB API (batched)}
\usage{
get_polymer_info(assembly_ids, batch_size = 200)
}
\arguments{
\item{assembly_ids}{A character vector of assembly IDs.}

\item{batch_size}{An integer specifying the number of assembly IDs to process in one batch (default: 400).}
}
\value{
A data frame with columns:
\itemize{
\item \code{entry_id}: The PDB entry ID.
\item \code{assembly_id}: The corresponding assembly ID.
\item \code{auth_asym_id}: The author-specified asymmetry ID.
\item \code{rcsb_id}: The RCSB polymer entity instance ID.
\item \code{title}: The title of the PDB entry.
\item \code{pfam_description}: The Pfam description for the entity.
}
}
\description{
This function queries the RCSB GraphQL API to retrieve polymer entity identifiers
(auth_asym_id and rcsb_id) along with Pfam descriptions for a given assembly ID.
}
