% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPfamAnnotationForPDBFilenames.R
\name{get_auth_asym_ids}
\alias{get_auth_asym_ids}
\title{Retrieve Polymer Entity Identifiers from RCSB API}
\usage{
get_auth_asym_ids(assembly_ids)
}
\arguments{
\item{assembly_ids}{A character vector of assembly IDs.}
}
\value{
A data frame with columns:
\itemize{
\item \code{entry_id}: The PDB entry ID.
\item \code{assembly_id}: The corresponding assembly ID.
\item \code{auth_asym_id}: The author-specified asymmetry ID.
\item \code{rcsb_id}: The RCSB polymer entity instance ID.
}
}
\description{
This function queries the RCSB GraphQL API to retrieve polymer entity identifiers
(auth_asym_id and rcsb_id) for a given assembly ID.
}
\keyword{:}
\keyword{internal}
