% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPfamAnnotationForPDBFilenames.R
\name{get_pfam_annotation}
\alias{get_pfam_annotation}
\title{Retrieve Pfam Annotations from RCSB API (batched)}
\usage{
get_pfam_annotation(rcsb_ids, batch_size = 1000)
}
\arguments{
\item{rcsb_ids}{A character vector of RCSB polymer entity instance IDs.}

\item{batch_size}{An integer specifying the number of entity IDs to process in one batch (default: 1000).}
}
\value{
A data frame with columns:
\itemize{
\item \code{rcsb_id}: The RCSB polymer entity instance ID.
\item \code{title}: The title of the PDB entry.
\item \code{pfam_description}: The Pfam description for the entity.
}
}
\description{
This function queries the RCSB GraphQL API to retrieve Pfam domain descriptions
for a given set of RCSB polymer entity instance IDs.
}
\keyword{:}
\keyword{internal}
